Updated Guessing Game Python code.
Added Welcome to User as well as asking for user's Name. Wrote comments as well as changed the whole structure to a double while loop. I as well changed the # of guesses, as have the code display a error message if the user inputs a character, Example a b c d e f g ... It doesn't count as a turn if entered a character. Again. The number will be randomly generated from 1 through 20.